Renaming all phase files for GASP
